+++ build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 	kotlin("plugin.spring") version "1.7.22"
 }
 
-group = "talk.message"
+group = "talk.messageService"
 version = "0.0.1-SNAPSHOT"
 java.sourceCompatibility = JavaVersion.VERSION_17
