+++ src/main/kotlin/talk/messageService/chatMessage/ChatMessageVM.kt
@@ -1,4 +1,4 @@
-package talk.message.service.chatMessage
+package talk.messageService.chatMessage
 
 data class ChatMessageVM(
         val name: String,
