+++ src/test/kotlin/talk/messageService/chat/ChatMessageResourceTest.kt
@@ -1,4 +1,4 @@
-package talk.message.service.chatMessage
+package talk.messageService.chat
 
 import io.kotest.core.spec.style.BehaviorSpec
 import kotlinx.coroutines.flow.collect
@@ -9,6 +9,7 @@ import org.springframework.messaging.rsocket.dataWithType
 import org.springframework.messaging.rsocket.retrieveFlow
 import org.springframework.messaging.rsocket.retrieveFlux
 import reactor.test.StepVerifier
+import talk.messageService.chatMessage.ChatMessageVM
 import java.time.Duration
 
 @SpringBootTest
@@ -17,7 +18,7 @@ class ChatMessageResourceTest(
 ) : BehaviorSpec({
     Given("chat 메세지 발신, 수신 stream 확인하기") {
 
-        fun chatStream() = rSocketRequester.route("stream.chats")
+        fun chatStream() = rSocketRequester.route("stream.chat.message")
 
         When("chat 메세지 정상 발송한 경우") {
             Then("1개의 메세지가 정상 수신된다") {
